modify move_speed change double

+++ src/CtrlS/UpgradeManager.java
@@ -129,9 +129,9 @@ public final class UpgradeManager {
      * @return The current movement speed.
      * @throws IOException In case of loading problems.
      */
-    public int getMovementSpeed() throws IOException {
+    public double getMovementSpeed() throws IOException {
         Properties properties = fileManager.loadUpgradeStatus();
-        return Integer.parseInt(properties.getProperty(MOVEMENT_SPEED, "1"));
+        return Double.parseDouble(properties.getProperty(MOVEMENT_SPEED, "1"));
     }
 
     /**
@@ -141,10 +141,10 @@ public final class UpgradeManager {
      * @throws IOException In case of saving problems.
      */
     public void addMovementSpeed() throws IOException {
-        int currentValue = getMovementSpeed();
+        double currentValue = getMovementSpeed();
         currentValue += shipStatus.getSpeedIn();
         Properties properties = fileManager.loadUpgradeStatus();
-        properties.setProperty(MOVEMENT_SPEED, Integer.toString(currentValue));
+        properties.setProperty(MOVEMENT_SPEED, Double.toString(currentValue));
         fileManager.saveUpgradeStatus(properties);
     }
