+++ src/entity/Ship.java
@@ -193,7 +193,7 @@ public class Ship extends Entity {
 	 *
 	 * @return Speed of the ship.
 	 */
-	public final int getSpeed() {
+	public final double getSpeed() {
 		return growth.getMoveSpeed();
 	}
